% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggshrug.R
\name{ggshrug}
\alias{ggshrug}
\title{Shruggie plot}
\usage{
ggshrug(df, x, y, ymin, ymax, xlab = "group", ylab = "quantity")
}
\arguments{
\item{df}{data frame containing parameters}

\item{x}{grouping variable}

\item{y}{measurement variable}

\item{ymin}{minimum y value (lower bound of shrug)}

\item{ymax}{maximum y value (upper bound of shrug)}
}
\value{
A ggplot object.
}
\description{
Creates a linerange plot formatted to look like a shrug emoticon.
}
\author{
Erin Jonaitis, \email{emjonaitis@gmail.com}
}
